// src/constants.d.ts
import { CompilerOptions } from './types';
export declare const rootName = "API";
export declare const bannerComment = "\n/**\n * This file was automatically generated by swagger-extract-definitions.\n * DO NOT MODIFY IT BY HAND.\n */\n\n";
export declare const cachedSchemas: {
    name: string;
    capitalized: string;
}[];
export declare const cachedSchemasWithBody: {
    name: string;
    capitalized: string;
}[];
export declare const methodsWithBody: string[];
export declare const isIgnoreHead: boolean, fileName: string;
export declare const outDir: string;
export declare const output: string;
export declare const targetFile: string;
export declare const compilerOptions: CompilerOptions;