// src/utils.d.ts
import { JSONSchema4 } from 'json-schema';
import { CompilerOptions } from './types';
export declare const capitalize: (str: string) => string;
export declare const normalizeTitle: (title: string) => string;
export declare const generateEndpointName: (endpoint: string) => string;
export declare const createSchemaObject: (title: string | undefined, properties?: Record<string, JSONSchema4>) => JSONSchema4;
export declare const patchCompilerOptions: (options: CompilerOptions, definitions: JSONSchema4) => CompilerOptions;
export declare const replaceRefs: (schema: JSONSchema4) => JSONSchema4;
export declare const validateSwaggerJson: (filePath: string) => any;
export declare const validateDirectory: (dir: string, name: string) => void;